+++ packages/web/app/src/components/ui/modal/index.ts
@@ -1,0 +1,9 @@
+export { ChangePermissionsModal } from './change-permissions';
+export { ConnectSchemaModal } from './connect-schema';
+export { CreateAccessTokenModal } from './create-access-token';
+export { CreateProjectModal } from './create-project';
+export { CreateTargetModal } from './create-target';
+export { DeleteOrganizationModal } from './delete-organization';
+export { DeleteProjectModal } from './delete-project';
+export { DeleteTargetModal } from './delete-target';
+export { TransferOrganizationOwnershipModal } from './transfer-organization-ownership';
